--- PicCharge/PicCharge/Service/SwiftDataStorage.swift
@@ -1,63 +1,0 @@
-//
-//  SwiftDataStorage.swift
-//  PicCharge
-//
-//  Created by 남유성 on 12/31/24.
-//
-
-import Foundation
-import SwiftData
-
-final class SwiftDataStorage<T: PersistentModel>: LocalStorage {
-    
-    private var container: ModelContainer
-    
-    init(container: ModelContainer) {
-        self.container = container
-    }
-}
-
-extension SwiftDataStorage {
-    func create(_ item: T) throws {
-        let context = ModelContext(container)
-        context.insert(item)
-        try context.save()
-    }
-    
-    func create(_ items: [T]) throws {
-        let context = ModelContext(container)
-        for item in items {
-            context.insert(item)
-        }
-        try context.save()
-    }
-    
-    func read(predicate: Predicate<T>? = nil,
-              sortDescriptors: SortDescriptor<T>...) throws -> [T] {
-        let context = ModelContext(container)
-        
-        let fetchDescriptor = FetchDescriptor<T>(
-            predicate: predicate,
-            sortBy: sortDescriptors
-        )
-        
-        return try context.fetch(fetchDescriptor)
-    }
-    
-    func update(_ item: T) throws {
-        let context = ModelContext(container)
-        try context.save()
-    }
-    
-    func delete(_ item: T) throws {
-        let context = ModelContext(container)
-        context.delete(item)
-        try context.save()
-    }
-    
-    func delete(where predicate: Predicate<T>) throws {
-        let context = ModelContext(container)
-        try context.delete(model: T.self, where: predicate)
-        try context.save()
-    }
-}
