+++ PicCharge/UserViewModel.swift
@@ -16,13 +16,13 @@ final class UserViewModel {
     private(set) var state: State = .init()
     
     @ObservationIgnored
-    private let localStorageService: LocalStorageService
+    private let localStorageService: LocalDataRepository
     
     @ObservationIgnored
     private let remoteStorageService: RemoteStorageService
     
     init(
-        localStorageService: LocalStorageService,
+        localStorageService: LocalDataRepository,
         remoteStorageService: RemoteStorageService
     ) {
         self.localStorageService = localStorageService
